Just before adding cache

+++ app/Http/Controllers/PostController.php
@@ -4,12 +4,27 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StorePost;
 use App\Models\BlogPost;
+use App\Models\User;
 use Illuminate\Contracts\Session\Session;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth')
+            ->only([
+                'create',
+                'store',
+                'edit',
+                'update',
+                'destroy',
+            ]);        
+    }
 
     /**
      * Display a listing of the resource.
@@ -31,9 +46,23 @@ class PostController extends Controller
         // }
 
         // dd(DB::getQueryLog());
-
         // dd(BlogPost::withCount('comments')->get());
-        return view('posts.index', ['posts' => BlogPost::withCount('comments')->get()]);
+        return view('posts.index', [
+            'posts' => BlogPost::latest()
+                ->withCount('comments')
+                ->with('user')
+                ->get(),
+            'mostCommented' => BlogPost::mostCommented()
+                ->take(5)
+                ->get(),
+            'mostActive' => User::withMostBlogPosts()
+                ->take(5)
+                ->get(),
+            'mostActiveLastMonth' => User::withMostBlogPostsLastMonth()
+                ->take(5)
+                ->get(),
+        ]);
+        // return view('posts.index', ['posts' => BlogPost::mostCommented()->withCount('comments')->get()]);
     }
 
     /**
@@ -43,6 +72,7 @@ class PostController extends Controller
      */
     public function create()
     {
+        $this->authorize(BlogPost::class);
         return view('posts.create');
     }
 
@@ -54,8 +84,13 @@ class PostController extends Controller
      */
     public function store(StorePost $request)
     {
+        $this->authorize(BlogPost::class);
+
         $validated = $request->validated();
 
+        $validated['user_id'] = $request->user()->id;
+
+        // $post = Auth::user()->blogPosts()->create($validated);
         $post = BlogPost::create($validated);
     
         // $post = new BlogPost();
@@ -76,6 +111,12 @@ class PostController extends Controller
      */
     public function show($id)
     {
+        // return view('posts.show', [
+        //     'post' => BlogPost::with(['comments' => function ($query) {
+        //         return $query->latest();
+        //     }])->findOrFail($id),
+        // ]);
+
         return view('posts.show', ['post' => BlogPost::with('comments')->findOrFail($id)]);
     }
 
@@ -87,7 +128,15 @@ class PostController extends Controller
      */
     public function edit($id)
     {
-        return view('posts.edit', ['post' => BlogPost::findOrFail($id)]);
+
+        $post = BlogPost::findOrFail($id);
+
+        $this->authorize($post);
+        // if (Gate::denies('update-post', $post)) {
+        //     abort(403, "You can't edit this blog post.");
+        // }
+
+        return view('posts.edit', ['post' => $post]);
     }
 
     /**
@@ -100,6 +149,12 @@ class PostController extends Controller
     public function update(StorePost $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+        $this->authorize($post);
+        // if (Gate::denies('update-post', $post)) {
+        //     abort(403, "You can't edit this blog post.");
+        // }
+
         $validated = $request->validated();
         $post->fill($validated);
         $post->save();
@@ -118,6 +173,13 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = BlogPost::findOrFail($id);
+
+        $this->authorize($post);
+
+        // if (Gate::denies('delete-post', $post)) {
+        //     abort(403, "You can't delete this blog post.");
+        // }
+
         $post->delete();
 
         session()->flash('status', 'The blog post was deleted successfully');
